Manually set gitops formula to 0.7.0
So the good news is, the automation is now able to push updates for
gitops, confirming weaveworks/corp#2423 is actually resolved. The
bad news is, it shouldn't push pre-releases.

I've manually set the URL and sha256's from the checksums in
git_checksums from
https://github.com/weaveworks/weave-gitops/releases/tag/v0.7.0
I've tested that it works on linux/x86_64 because that's what I can
access.